Optimize the decoding of generic http2
